# api/swagger/admin.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hotels API",
    "description": "admin site",
    "version": "v1"
  },
  "servers": [
    {
    }
  ],
  "paths": {
...
    "/api/admin/hotels": {
      "get": {
        "summary": "Get hotels",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "Default: 1",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "description": "Default: 20",
            "schema": {
              "type": "string",
              "example": "20"
            }
          },
          {
            "name": "verified",
            "in": "path",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/getHotel"
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create hotel",
        "tags": [
          "Hotels"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postHotel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getHotel"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/hotels/{id}": {
      "get": {
        "summary": "Get hotel",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getHotel"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update hotel",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postHotel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getHotel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete hotel",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Hotel was successfully deleted"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/hotels/{id}/actions/markAsVerified": {
      "put": {
        "summary": "Mark Hotel as verified",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getHotel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "getUser": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "634e8652d6389a1eb1268862"
          },
          "email": {
            "type": "string",
            "example": "example@gmail.com"
          },
          "firstName": {
            "type": "string",
            "example": "Ken"
          },
          "lastName": {
            "type": "string",
            "example": "Adams"
          },
          "phone": {
            "type": "string",
            "example": "380977777777"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "owner", "guest"],
            "example": "guest"
          }
        }
      },
...
      "postHotel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "getHotel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "634e8652d6389a1eb1268862"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/getUser"
          },
          "description": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/getAsset"
            }
          }
        }
      },
      "getAsset": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          }
        }
      }
    }
  }
}